% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harmonize.R
\name{apply_intervention_map}
\alias{apply_intervention_map}
\title{Apply an intervention mapping to a dataset}
\usage{
apply_intervention_map(data, intervention_col, map)
}
\arguments{
\item{data}{data.frame}

\item{intervention_col}{name of the intervention column (string)}

\item{map}{tibble with columns original and suggested (as returned by suggest_intervention_map)}
}
\value{
data.frame with intervention values harmonized (and attribute metaclean_fixes updated)
}
\description{
Apply an intervention mapping to a dataset
}
