% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validity_checks.R
\name{check_invalid_values}
\alias{check_invalid_values}
\title{Check for common invalid data entries (events > n, decimals where integers expected)}
\usage{
check_invalid_values(df, values)
}
\arguments{
\item{df}{data.frame}

\item{values}{list with names 'n' and optionally 'events'}
}
\value{
list of tibbles with problems
}
\description{
Check for common invalid data entries (events > n, decimals where integers expected)
}
