% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/string_cleaning.R
\name{clean_core_strings}
\alias{clean_core_strings}
\title{Clean string columns: trim, unify case, remove punctuation (keeps alnum + spaces), unify spaces to _}
\usage{
clean_core_strings(df, cols)
}
\arguments{
\item{df}{data.frame}

\item{cols}{character vector of column names to clean}
}
\value{
data.frame
}
\description{
Clean string columns: trim, unify case, remove punctuation (keeps alnum + spaces), unify spaces to _
}
