% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intervention_variants.R
\name{detect_name_variants}
\alias{detect_name_variants}
\title{Detect likely variants in intervention labels using Jaro-Winkler distance + clustering}
\usage{
detect_name_variants(x, threshold = 0.15)
}
\arguments{
\item{x}{character vector of intervention labels}

\item{threshold}{numeric (0-1) height to cut tree; lower -> stricter grouping}
}
\value{
tibble with name and cluster id
}
\description{
Detect likely variants in intervention labels using Jaro-Winkler distance + clustering
}
