% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imbalance_outlier.R
\name{detect_outliers_and_imbalance}
\alias{detect_outliers_and_imbalance}
\title{Detect outliers in numeric columns using IQR and MAD and detect arm imbalance}
\usage{
detect_outliers_and_imbalance(
  df,
  n_col = "n",
  mean_col = NULL,
  sd_col = NULL,
  imbalance_threshold = 10
)
}
\arguments{
\item{df}{data.frame}

\item{n_col}{name of sample size column}

\item{mean_col}{mean column name (optional)}

\item{sd_col}{sd column name (optional)}

\item{imbalance_threshold}{numeric ratio (max/min) to flag imbalance (default 10)}
}
\value{
list with outlier tibbles and imbalance tibble
}
\description{
Detect outliers in numeric columns using IQR and MAD and detect arm imbalance
}
