% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outliers.R
\name{detect_outliers_meta}
\alias{detect_outliers_meta}
\title{Detect outliers in sample size (n) using robust IQR method}
\usage{
detect_outliers_meta(df, n_col)
}
\arguments{
\item{df}{data.frame}

\item{n_col}{column name (string) for sample size}
}
\value{
tibble of rows flagged as outliers
}
\description{
Detect outliers in sample size (n) using robust IQR method
}
