% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nma_checks.R
\name{detect_subnetworks}
\alias{detect_subnetworks}
\title{Detect connected components (subnetworks) for NMA using igraph}
\usage{
detect_subnetworks(df, study_id, intervention)
}
\arguments{
\item{df}{data.frame (long format)}

\item{study_id}{column name for study id}

\item{intervention}{column name for interventions}
}
\value{
igraph components object (sizes, membership) or list with components
}
\description{
Detect connected components (subnetworks) for NMA using igraph
}
