% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.metaclean_report}
\alias{print.metaclean_report}
\title{Print method for metaclean reports}
\usage{
\method{print}{metaclean_report}(x, ...)
}
\arguments{
\item{x}{object returned by metaclean}
}
\description{
Print method for metaclean reports
}
